Split inlined implementations into its own header.
